Squelched the warning about new Git defaults
Git's standard behaviour regarding incomplete push commands to date is
to only update remote branches that match the names of local ones. The
push will proceed for all branches this is true, and fail for any
branches that don't exist by the same name on the remote.

For example: someone makes a branch available on the remote `github`
called `feature1` that you pull and rename locally to `add-menu`. When
invoking any of the following commands, the push will not result in
the (likely) intended effect of modifying the `feature1` branch on the
remote.

	git push github add-menu
	git push github
	git push
	git push github HEAD

This is unlike the behaviour of `pull`, which *will* pull changes from
`feature1` into the local `add-menu` branch. The correct way to
address this branch name mismatch is to

	git push github add-menu:feature1

...which is nice and explicit, but potentially confusing when
everything updates as expected except for `feature1` when you omit
this.

Git's default behaviour is in the `push.default` setting, the value of
which is `matching` in recent versions. With Git 2.0, this default
will change to `simple`, which will explicitly fail the push with a
message telling you to use the explicit command.

Simply put, Git is warning about this now on commit, and this change
adopts the new behaviour.

Reference:
http://longair.net/blog/2011/02/27/an-asymmetry-between-git-pull-and-git-push/