feat(macos): switch from terminal.app to wezterm
Terminal.app doesn't support true colour; this is a showstopper for
neovim. kitty, alacritty and wezterm are all gpu-accelerated open source
alternatives and wezterm wins the trials for me.

Wezterm is proactively maintained, works well and has synergy with
lua-based neovim configuration.

Homebrew based install, configuration file included in this commit with
a tentative colour scheme choice (this one results in a more subtle
starship prompt colourisation and is easy on the eyes akin to my older
monochromatic custom prompt). Other colour schemes trialled are mostly
the usual suspects but left in the configuration commented out for
reference.

The existing terminal shortcut configured in karabiner has been changed
to:

  - <Ctrl+Cmd/Alt+T> open new wezterm tab in existing instance
  - <Ctrl+Cmd/Alt+Shift+T> open new wezterm window in existing instance