feat(shell): apply mr to repo subsets via wrapper
myrepos lacks a native way to operate on subsets of repositories. This
is often desirable, and can be done via careful modularisation of the mr
configuration.

This wrapper, `ms`, supplies an additional argument to `mr` which seeds
fzf, so you can select a specific config file and then operate on only
the repos defined within.

It's simple, but has worked well for relatively extensive use so far.