feat(py): use conda by default
pyenv is annoying and doesn't work how I'm used to. It's too
presumptuous and seems to dominate the environment by default. I was
originally trying it out because I wanted to move towards letting poetry
handle everything rather than hook into mamba/conda environments. The
option for this seemed to be using pyenv (without venv). This works, but
poetry depends on that pyenv python binary anyway, so it's nowhere near
as self contained as conda.

Solution
---

By default, don't set up pyenv outside of a pyenv-enable command, and
add a function to disable it (pyenv-disable).

Remains to be seen how useful this is but currently just prefer poetry
with mamba/conda.